+++ CHANGELOG.md
@@ -74,3 +74,16 @@ setters and getters
 - Implemented common methods in TaskManagerSecondary
 - Implemented common methods in Task
 - Implemented iterator for TaskManagerOnSequence
+
+## 2026.04.22
+
+### Added
+
+- Wrote JUnit tests for the kernel and secondary methods for TaskManagerOnSequence
+- Created a use case of a daily planner
+
+## 2026.04.23
+
+### Added
+
+- Created a use case of a school course work organizer